알람 부분 이랑 일단 toggle만 전역으로
